python-pip-requirements-dev.txt
pytest==4.3.1
pytest-django==3.4.8